+++ localization/EN_US.json
@@ -1,0 +1,43 @@
+﻿{
+	"pets": {
+		"cryogen": "Cryogen",
+		"clamClone": "Forgotten Shadow of Calamitas",
+		"aquaticScoog": "Aquatic Scourge",
+		"deus": "Astrum Deus",
+		"clover": "Clover",
+		"aureus": "Astrum Aureus",
+		"brimmy": "Brimstone Elemental",
+		"birb": "Dragonfolly",
+		"ceaseless": "Ceaseless Void",
+		"crabulon": "Crabulon",
+		"desertScoog": "Desert Scourge",
+		"DoG": "DoG",
+		"ares": "Ares",
+		"apollo": "Apollo",
+		"artemis": "Artemis",
+		"thanatos": "Thanatos",
+		"drae": "Draedon",
+		"exoMayhem": "Exo Mechahem",
+		"hive": "Hive Mind",
+		"nahita": "Anahita",
+		"leviathan": "Leviathan",
+		"oldDuke": "Old Duke",
+		"perf": "Perforators",
+		"golia": "PBG",
+		"polter": "Polterghast",
+		"18plusWyrm": "18+ wyrm",
+		"prov": "Providence",
+		"signus": "Signus",
+		"ravager": "Ravager",
+		"slimegod": "Slime God",
+		"weebaa": "Storm Weaver",
+		"eoc": "Eye o Cthulhu",
+		"clam": "Calamitas",
+		"yharon": "Yharon",
+		"sepulcher": "Sepulcher",
+		"wulfrumDevourer": "Wulfrum worm",
+		"goozma": "Goozma",
+		"kingSlime": "King Slime",
+		"queenBee": "Queen bee"
+	}
+}
