+++ localization/ZH_APRILFOOL.json
@@ -1,0 +1,43 @@
+﻿{
+	"pets": {
+		"cryogen": "冰生",
+		"clamClone": "灾厄米塔斯克隆",
+		"aquaticScoog": "海洋鞭子",
+		"deus": "星虫",
+		"clover": "Clover",
+		"aureus": "黄金之星",
+		"brimmy": "硫磺火元素",
+		"birb": "愚龙",
+		"ceaseless": "无尽虚空",
+		"crabulon": "菌生蟹",
+		"desertScoog": "荒漠鞭子",
+		"DoG": "狗",
+		"ares": "阿端斯",
+		"apollo": "阿波罗",
+		"artemis": "阿尔忒弥斯",
+		"thanatos": "塔纳托斯",
+		"drae": "登",
+		"exoMayhem": "异星机甲 全",
+		"hive": "蜂巢思维",
+		"nahita": "纳西塔",
+		"leviathan": "李伟坦",
+		"oldDuke": "不存在的东西",
+		"perf": "穿刺",
+		"golia": "疫使歌利亚",
+		"polter": "波特灵骚",
+		"18plusWyrm": "18+蛟蜃",
+		"prov": "猥亵天神",
+		"signus": "西格纳斯",
+		"ravager": "劫掠兽",
+		"slimegod": "史莱姆之神",
+		"weebaa": "风暴织网者",
+		"eoc": "克苏鲁之眼",
+		"clam": "至尊灾厄米塔斯",
+		"yharon": "鸭绒",
+		"sepulcher": "九月",
+		"wulfrumDevourer": "钨钢吞噬者",
+		"goozma": "古兹马",
+		"kingSlime": "史莱姆王",
+		"queenBee": "蜂王"
+	}
+}
