+++ localization/ZH_HANS.json
@@ -1,0 +1,43 @@
+﻿{
+	"pets": {
+		"cryogen": "极地冰灵",
+		"clamClone": "灾厄之影",
+		"aquaticScoog": "渊海灾虫",
+		"deus": "星神游龙",
+		"clover": "Clover",
+		"aureus": "白金星舰",
+		"brimmy": "硫磺火元素",
+		"birb": "痴愚金龙",
+		"ceaseless": "无尽虚空",
+		"crabulon": "菌生蟹",
+		"desertScoog": "荒漠灾虫",
+		"DoG": "神明吞噬者",
+		"ares": "XF09-阿瑞斯",
+		"apollo": "阿波罗",
+		"artemis": "阿尔忒弥斯",
+		"thanatos": "塔纳托斯",
+		"drae": "嘉登",
+		"exoMayhem": "星流巨械 全",
+		"hive": "腐巢意志",
+		"nahita": "阿纳西塔",
+		"leviathan": "利维坦",
+		"oldDuke": "硫海遗爵",
+		"perf": "血肉宿主",
+		"golia": "歌莉娅",
+		"polter": "噬魂幽花",
+		"18plusWyrm": "始源妖龙",
+		"prov": "亵渎天神",
+		"signus": "西格纳斯",
+		"ravager": "毁灭魔像",
+		"slimegod": "史莱姆之神",
+		"weebaa": "风暴编织者",
+		"eoc": "克苏鲁之眼",
+		"clam": "至尊灾厄",
+		"yharon": "犽戎",
+		"sepulcher": "灾坟魔物",
+		"wulfrumDevourer": "钨钢吞噬者",
+		"goozma": "古泽玛",
+		"kingSlime": "史莱姆王",
+		"queenBee": "蜂王"
+	}
+}
